Move ref.Val -> api.expr.Value converter to provider.go
