+++ server/BUILD.bazel
@@ -16,7 +16,6 @@ go_library(
         "//common:go_default_library",
         "//common/types:go_default_library",
         "//common/types/ref:go_default_library",
-        "//common/types/traits:go_default_library",
         "@com_google_cel_spec//proto/test/v1/proto2:test_all_types_go_proto",
         "@com_google_cel_spec//proto/test/v1/proto3:test_all_types_go_proto",
         "@org_golang_google_genproto//googleapis/api/expr/conformance/v1alpha1:go_default_library",
@@ -26,8 +25,6 @@ go_library(
         "@org_golang_google_grpc//status:go_default_library",
         "@org_golang_google_protobuf//proto:go_default_library",
         "@org_golang_google_protobuf//types/known/anypb:go_default_library",
-        "@org_golang_google_protobuf//types/known/durationpb:go_default_library",
-        "@org_golang_google_protobuf//types/known/timestamppb:go_default_library",
     ],
 )
